+++ datastore/event_delivery.go
@@ -166,7 +166,7 @@ func (db *eventDeliveryRepo) UpdateEventDeliveryWithAttempt(ctx context.Context,
 	return nil
 }
 
-func (db *eventDeliveryRepo) LoadEventDeliveriesPaged(ctx context.Context, groupID, appID, eventID string, status convoy.EventDeliveryStatus, searchParams models.SearchParams, pageable models.Pageable) ([]convoy.EventDelivery, pager.PaginationData, error) {
+func (db *eventDeliveryRepo) LoadEventDeliveriesPaged(ctx context.Context, groupID, appID, eventID string, status []convoy.EventDeliveryStatus, searchParams models.SearchParams, pageable models.Pageable) ([]convoy.EventDelivery, pager.PaginationData, error) {
 	filter := bson.M{
 		"document_status": bson.M{"$ne": convoy.DeletedDocumentStatus},
 		"created_at":      getCreatedDateFilter(searchParams),
@@ -175,7 +175,7 @@ func (db *eventDeliveryRepo) LoadEventDeliveriesPaged(ctx context.Context, group
 	hasAppFilter := !util.IsStringEmpty(appID)
 	hasGroupFilter := !util.IsStringEmpty(groupID)
 	hasEventFilter := !util.IsStringEmpty(eventID)
-	hasStatusFilter := !util.IsStringEmpty(string(status))
+	hasStatusFilter := len(status) > 0
 
 	if hasAppFilter {
 		filter["app_metadata.uid"] = appID
@@ -190,7 +190,7 @@ func (db *eventDeliveryRepo) LoadEventDeliveriesPaged(ctx context.Context, group
 	}
 
 	if hasStatusFilter {
-		filter["status"] = status
+		filter["status"] = bson.M{"$in": status}
 	}
 
 	var eventDeliveries []convoy.EventDelivery