+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.3.0
+* [Bugfix] Fixed hash computation bug #269
+* [Bugfix] Bundled UI into binary builds. #273
+* [Bugfix] Don't enqueue discarded events #253
+* [Enhancement] Build Convoy dashboard UI into npm package in `dashboard-component` #272
+
+## 0.3.0-rc
+* [Feature] URL/Events: User can specify events to each endpoint.
+* [Feature] Batch Retries: User can batch retry multiple event deliveries at once.
+* [Feature] Unified auth configuration for both UI and API.
+* [Feature] Added minimal rbac for super user, admin and ui admin.
+* [Feature] New tab to view event deliveries for events. 
+* [Feature] Filter event deliveries by delivery status, app and date range
+* [Feature] View event deliveries status for each event from event's details section. 
+* [Feature] Introduced Groups: To support multi-tenancy for multiple products to pipe events as separete groups.
+* [Feature] Persist events and event deliveries filters, active group and active logs tab with page reload.
+
 ## 0.2.0
 * [Feature] Add disable events and send email notifications.
 * [Feature] Re-activate endpoints by re-trying a non-successful event.